requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.5
django-allauth==0.61.1
django-cors-headers==4.5.0
django-redis==5.4.0
djangorestframework==3.15.2
elastic-transport==8.13.0
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
google-api-core==2.21.0
google-api-python-client==2.130.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.65.0
grpcio==1.64.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
kiwisolver==1.4.5
msgpack==1.1.0
mysqlclient==2.2.4
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
redis==5.1.1
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.0
service-identity==24.2.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.2
threadpoolctl==3.5.0
tomli==2.2.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
zope.interface==7.2
